refactor: moved func_expression function
